Reduce tightness of coupling in NodeController
