+++ pkg/controller/node/nodecontroller_test.go
@@ -555,14 +555,14 @@ func TestMonitorNodeStatusEvictPods(t *testing.T) {
 		}
 
 		nodeController.podEvictor.Try(func(value TimedValue) (bool, time.Duration) {
-			remaining, _ := nodeController.deletePods(value.Value)
+			remaining, _ := deletePods(item.fakeNodeHandler, nodeController.recorder, value.Value, nodeController.daemonSetStore)
 			if remaining {
 				nodeController.terminationEvictor.Add(value.Value)
 			}
 			return true, 0
 		})
 		nodeController.podEvictor.Try(func(value TimedValue) (bool, time.Duration) {
-			nodeController.terminatePods(value.Value, value.AddedAt)
+			terminatePods(item.fakeNodeHandler, nodeController.recorder, value.Value, value.AddedAt, nodeController.maximumGracePeriod)
 			return true, 0
 		})
 		podEvicted := false
@@ -1082,7 +1082,7 @@ func TestNodeDeletion(t *testing.T) {
 		t.Errorf("unexpected error: %v", err)
 	}
 	nodeController.podEvictor.Try(func(value TimedValue) (bool, time.Duration) {
-		nodeController.deletePods(value.Value)
+		deletePods(fakeNodeHandler, nodeController.recorder, value.Value, nodeController.daemonSetStore)
 		return true, 0
 	})
 	podEvicted := false
@@ -1220,12 +1220,12 @@ func TestCheckPod(t *testing.T) {
 
 	for i, tc := range tcs {
 		var deleteCalls int
-		nc.forcefullyDeletePod = func(_ *api.Pod) error {
+		forcefullyDeletePodsFunc := func(_ *api.Pod) error {
 			deleteCalls++
 			return nil
 		}
 
-		nc.maybeDeleteTerminatingPod(&tc.pod)
+		nc.maybeDeleteTerminatingPod(&tc.pod, nc.nodeStore.Store, forcefullyDeletePodsFunc)
 
 		if tc.prune && deleteCalls != 1 {
 			t.Errorf("[%v] expected number of delete calls to be 1 but got %v", i, deleteCalls)
@@ -1237,17 +1237,7 @@ func TestCheckPod(t *testing.T) {
 }
 
 func TestCleanupOrphanedPods(t *testing.T) {
-	newPod := func(name, node string) api.Pod {
-		return api.Pod{
-			ObjectMeta: api.ObjectMeta{
-				Name: name,
-			},
-			Spec: api.PodSpec{
-				NodeName: node,
-			},
-		}
-	}
-	pods := []api.Pod{
+	pods := []*api.Pod{
 		newPod("a", "foo"),
 		newPod("b", "bar"),
 		newPod("c", "gone"),
@@ -1263,12 +1253,12 @@ func TestCleanupOrphanedPods(t *testing.T) {
 
 	var deleteCalls int
 	var deletedPodName string
-	nc.forcefullyDeletePod = func(p *api.Pod) error {
+	forcefullyDeletePodFunc := func(p *api.Pod) error {
 		deleteCalls++
 		deletedPodName = p.ObjectMeta.Name
 		return nil
 	}
-	nc.cleanupOrphanedPods()
+	cleanupOrphanedPods(pods, nc.nodeStore.Store, forcefullyDeletePodFunc)
 
 	if deleteCalls != 1 {
 		t.Fatalf("expected one delete, got: %v", deleteCalls)
